Add authentication and authorisation to Ibutsu
- Add a login page
- Add a login controller
- Use JWT authentication
- Make HttpClient service object that transparently auths requests
- Add OAuth2 login to Ibutsu
- Add tests for the login controller
- Add Keycloak integration
- Add authorisation, users can only see resources for projects they are part of
- Use the 'from_file' method if it exists
- Add support for running the dev servers under TLS

+++ backend/ibutsu_server/controllers/dashboard_controller.py
@@ -2,10 +2,11 @@ import connexion
 from ibutsu_server.db.base import session
 from ibutsu_server.db.models import Dashboard
 from ibutsu_server.db.models import WidgetConfig
+from ibutsu_server.util.projects import project_has_user
 from ibutsu_server.util.uuid import validate_uuid
 
 
-def add_dashboard(dashboard=None):
+def add_dashboard(dashboard=None, token_info=None, user=None):
     """Create a dashboard
 
     :param body: Dashboard
@@ -16,13 +17,15 @@ def add_dashboard(dashboard=None):
     if not connexion.request.is_json:
         return "Bad request, JSON required", 400
     dashboard = Dashboard.from_dict(**connexion.request.get_json())
+    if dashboard.project_id and not project_has_user(dashboard.project_id, user):
+        return "Forbidden", 403
     session.add(dashboard)
     session.commit()
     return dashboard.to_dict(), 201
 
 
 @validate_uuid
-def get_dashboard(id_):
+def get_dashboard(id_, token_info=None, user=None):
     """Get a single dashboard by ID
 
     :param id: ID of test dashboard
@@ -30,15 +33,15 @@ def get_dashboard(id_):
 
     :rtype: Dashboard
     """
-    dashboard = Dashboard.query.filter(Dashboard.id == id_).first()
-    if not dashboard:
-        dashboard = Dashboard.query.get(id_)
+    dashboard = Dashboard.query.get(id_)
     if not dashboard:
         return "Dashboard not found", 404
+    if dashboard and dashboard.project and not project_has_user(dashboard.project, user):
+        return "Forbidden", 403
     return dashboard.to_dict()
 
 
-def get_dashboard_list(project_id=None, user_id=None, page=1, page_size=25):
+def get_dashboard_list(project_id=None, page=1, page_size=25, token_info=None, user=None):
     """Get a list of dashboards
 
     :param project_id: Filter dashboards by project ID
@@ -53,10 +56,14 @@ def get_dashboard_list(project_id=None, user_id=None, page=1, page_size=25):
     :rtype: DashboardList
     """
     query = Dashboard.query
+    if "project_id" in connexion.request.args:
+        project_id = connexion.request.args["project_id"]
     if project_id:
+        if not project_has_user(project_id, user):
+            return "Forbidden", 403
         query = query.filter(Dashboard.project_id == project_id)
-    if user_id:
-        query = query.filter(Dashboard.user_id == user_id)
+    if user:
+        query = query.filter(Dashboard.user_id == user)
     offset = (page * page_size) - page_size
     total_items = query.count()
     total_pages = (total_items // page_size) + (1 if total_items % page_size > 0 else 0)
@@ -72,7 +79,7 @@ def get_dashboard_list(project_id=None, user_id=None, page=1, page_size=25):
     }
 
 
-def update_dashboard(id_, dashboard=None):
+def update_dashboard(id_, dashboard=None, token_info=None, user=None):
     """Update a dashboard
 
     :param id: ID of test dashboard
@@ -84,16 +91,23 @@ def update_dashboard(id_, dashboard=None):
     """
     if not connexion.request.is_json:
         return "Bad request, JSON required", 400
+    dashboard_dict = connexion.request.get_json()
+    if dashboard_dict.get("metadata", {}).get("project") and not project_has_user(
+        dashboard_dict["metadata"]["project"], user
+    ):
+        return "Forbidden", 403
     dashboard = Dashboard.query.get(id_)
     if not dashboard:
         return "Dashboard not found", 404
+    if project_has_user(dashboard.project, user):
+        return "Forbidden", 403
     dashboard.update(connexion.request.get_json())
     session.add(dashboard)
     session.commit()
     return dashboard.to_dict()
 
 
-def delete_dashboard(id_):
+def delete_dashboard(id_, token_info=None, user=None):
     """Deletes a dashboard
 
     :param id: ID of the dashboard to delete
@@ -104,10 +118,11 @@ def delete_dashboard(id_):
     dashboard = Dashboard.query.get(id_)
     if not dashboard:
         return "Not Found", 404
-    else:
-        widget_configs = WidgetConfig.query.filter(WidgetConfig.dashboard_id == dashboard.id).all()
-        for widget_config in widget_configs:
-            session.delete(widget_config)
-        session.delete(dashboard)
-        session.commit()
-        return "OK", 200
+    if not project_has_user(dashboard.project, user):
+        return "Forbidden", 403
+    widget_configs = WidgetConfig.query.filter(WidgetConfig.dashboard_id == dashboard.id).all()
+    for widget_config in widget_configs:
+        session.delete(widget_config)
+    session.delete(dashboard)
+    session.commit()
+    return "OK", 200
